BackendDataProcessing/requirements.txt
APScheduler==3.3.1
cassandra-driver==3.9.0
contextlib2==0.5.4
dnet==1.12
Flask==0.12
Flask-PyMongo==0.4.1
flaskr==0.0.0
funcsigs==1.0.2
futures==3.1.1
googlefinance==0.7
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
kafka-python==1.3.3
MarkupSafe==1.0
pytz==2017.2
raven==6.0.0
requests==2.13.0
six==1.10.0
tzlocal==1.4
Werkzeug==0.12.1
